Move html highlightig to a separate module
